backend/requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-openai
supabase
python-dotenv
pydantic
sse-starlette
python-multipart
